refactor: Se modificó la ruta de los componentes que almacenan los iconos del sistema
