+++ src/commons/ui/icons/actionIcons/searchIco.vue
@@ -1,0 +1,14 @@
+<script setup>
+const props = defineProps({
+    color: {
+        type: String,
+        default: 'white'
+    }
+});
+</script>
+
+<template>
+<svg id="Capa_1" data-name="Capa 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 500 500" width="25" height="21">
+  <path :fill="props.color" class="cls-1" d="M326.82,365.01c-103.89,71.05-219.16,43.24-284.15-36.7C-17.87,253.83-13.33,136.56,52.08,66.44,123.44-10.06,232.14-19.09,313.14,32.09c78.47,49.57,134.3,174.51,51.93,295.29,22.55-3.71,33.93,11.09,46.61,23.85,25.64,25.8,51.5,51.38,77.13,77.2,14.73,14.83,14.89,25.19.68,40.42-7.23,7.75-14.75,15.27-22.52,22.49-12.07,11.21-24.56,11.78-36.31.28-31.78-31.09-62.94-62.8-94.5-94.11-8.8-8.73-8.34-20-9.34-32.5ZM327.5,204.65c.4-69.9-54.53-126.31-123.34-126.66-69.38-.35-125.55,54.22-125.83,122.24-.3,72.15,52.84,126.59,123.79,126.83,69.93.23,124.98-53.52,125.38-122.41Z"/>
+</svg>
+</template>
