+++ src/commons/ui/icons/index.js
@@ -1,9 +1,9 @@
-import buyIco from '@/commons/ui/icons/buyIco.vue';
-import catalogueIco from '@/commons/ui/icons/catalogueIco.vue';
-import organizationIco from '@/commons/ui/icons/organizationIco.vue';
-import sellIco from '@/commons/ui/icons/sellIco.vue';
-import settingsIco from '@/commons/ui/icons/settingsIco.vue';
-import logoutIco from '@/commons/ui/icons/logoutIco.vue';
+import buyIco from '@/commons/ui/icons/sidebarIcons/buyIco.vue';
+import catalogueIco from '@/commons/ui/icons/sidebarIcons/catalogueIco.vue';
+import organizationIco from '@/commons/ui/icons/sidebarIcons/organizationIco.vue';
+import sellIco from '@/commons/ui/icons/sidebarIcons/sellIco.vue';
+import settingsIco from '@/commons/ui/icons/sidebarIcons/settingsIco.vue';
+import logoutIco from '@/commons/ui/icons/sidebarIcons/logoutIco.vue';
 
 export {
   buyIco,
