+++ src/commons/ui/icons/sidebarIcons/buyIco.vue
