+++ src/commons/ui/icons/sidebarIcons/catalogueIco.vue
