+++ src/commons/ui/icons/sidebarIcons/logoutIco.vue
