+++ src/commons/ui/icons/sidebarIcons/organizationIco.vue
