+++ src/commons/ui/icons/sidebarIcons/sellIco.vue
