+++ src/commons/ui/icons/sidebarIcons/settingsIco.vue
